feat: 値オブジェクトの基底クラスではなく、Enumをconstで表現した形式で実装 #1
